Fixed allmost all bugs

+++ MinnesmarkEditor/src/mmMap/MmMapViewer.java
@@ -44,6 +44,15 @@ import java.util.HashSet;
 import java.util.Set;
 
 import javax.imageio.ImageIO;
+import javax.print.Doc;
+import javax.print.DocFlavor;
+import javax.print.DocPrintJob;
+import javax.print.PrintService;
+import javax.print.PrintServiceLookup;
+import javax.print.SimpleDoc;
+import javax.print.attribute.HashPrintRequestAttributeSet;
+import javax.print.attribute.PrintRequestAttributeSet;
+import javax.print.attribute.standard.Copies;
 import javax.swing.*;
 
 import mmAccordionMenu.*;
@@ -209,7 +218,9 @@ public class MmMapViewer extends JPanel implements Printable {
 		
 	JSONObject initializeEvents = new JSONObject();
 	JSONObject initializeActions = new JSONObject();
-	JSONArray action = new JSONArray();
+	JSONArray action;
+	
+	JSONArray startActionArray;
 	
 	//key:552b7cf0010662a89bcbd90b9727186c4fb97ffe7883e11a2ebba59765d7f461
 	
@@ -240,8 +251,11 @@ public class MmMapViewer extends JPanel implements Printable {
 	//events dialog
 	JDialog eventsDialog;
 
-	
+	//file open
+	boolean isFileOpen;
 
+
+	
 
 	public MmMapViewer()
 	{
@@ -314,7 +328,7 @@ public class MmMapViewer extends JPanel implements Printable {
 		mapKit.setCursor(new Cursor(Cursor.HAND_CURSOR));
 		((DefaultTileFactory)mapKit.getMainMap().getTileFactory()).setThreadPoolSize(8);
 		Toolkit toolkit = Toolkit.getDefaultToolkit();
-		mapKit.setPreferredSize(new Dimension(toolkit.getScreenSize().width-380,toolkit.getScreenSize().height-175));
+		mapKit.setPreferredSize(new Dimension(toolkit.getScreenSize().width-460,toolkit.getScreenSize().height-200));
 		add(mapKit);
 		mapKit.setAddressLocationShown(true);
 		mapKit.getMainMap().setPanEnabled(true);
@@ -1087,6 +1101,15 @@ public class MmMapViewer extends JPanel implements Printable {
 	
 	public JDialog getMapMarkersDialog() {
 		return mapMarkersDialog;
+	}
+
+	public boolean isFileOpen() {
+		return isFileOpen;
+	}
+
+
+	public void setFileOpen(boolean isFileOpen) {
+		this.isFileOpen = isFileOpen;
 	}
 
 
@@ -2038,7 +2061,8 @@ public class MmMapViewer extends JPanel implements Printable {
 	
 	public void createJSONFile(String fileName,String filePath,ArrayList<MmGlobalMarkerEvents> markerEvents,MmStartEvents startEvents)
 	{
-		
+	
+		action = new JSONArray();
 		JSONArray eventsArray = new JSONArray();
 		
 		JSONObject numCollectItems = new JSONObject();
@@ -2087,29 +2111,37 @@ public class MmMapViewer extends JPanel implements Printable {
 		
 		    eventsArray.put(stationAudioEvent.getAudioEvent());
 		
-		    MmAudioEvent swingAudioEvent = new MmAudioEvent();
-		    swingAudioEvent.setEventName("wayPointAudio");
-  	        swingAudioEvent.setAudioFileName("wayPoint.m4a");
-		    swingAudioEvent.setSourcePath(System.getProperty("user.dir")+"/audios/wayPoint.m4a");
-		    swingAudioEvent.setDestinationPath(filePath);
-	  	    swingAudioEvent.makeJSONObject();
-		    if(!swingAudioEvent.JSONActions())
+		    for(int i=0;i<events.getStations().size();i++)
 		    {	
-		    	JOptionPane.showMessageDialog(null, "Kan inte spara json filen");
-		    	File file = new File(filePath);
-		    	if(file.isDirectory())
-		    	{
-		    		File[] files = file.listFiles();
-		    		if(files.length>0)
-		    		{
-		    			files[0].delete();
-		    		}
-		    		file.delete();
-		    	}	
-		    	return;
-		    }	
+		       if(events.getStations().get(i).getStationType())
+		       {	   
+		          MmAudioEvent swingAudioEvent = new MmAudioEvent();
+		          swingAudioEvent.setEventName("wayPointAudio");
+  	              swingAudioEvent.setAudioFileName("wayPoint.m4a");
+		          swingAudioEvent.setSourcePath(System.getProperty("user.dir")+"/audios/wayPoint.m4a");
+		          swingAudioEvent.setDestinationPath(filePath);
+	  	          swingAudioEvent.makeJSONObject();
+		          if(!swingAudioEvent.JSONActions())
+		          {	
+		    	       JOptionPane.showMessageDialog(null, "Kan inte spara json filen");
+		    	       File file = new File(filePath);
+		    	       if(file.isDirectory())
+		    	       {
+		    		        File[] files = file.listFiles();
+		    		        if(files.length>0)
+		    		        {
+		    			         files[0].delete();
+		    		        }
+		    		       file.delete();
+		    	       }	
+		    	       return;
+		          }
+		          eventsArray.put(swingAudioEvent.getAudioEvent());
+		          break;
+		       }  
+		    }   
 		    
-		    eventsArray.put(swingAudioEvent.getAudioEvent());
+		    
 		    
 		    try
 		    {
@@ -2153,6 +2185,19 @@ public class MmMapViewer extends JPanel implements Printable {
 			   if(!writeMap3DObjects(filePath))
 			   {
 				   JOptionPane.showMessageDialog(null, "Kan inte spara json filen");
+				   
+				   File file = new File(filePath);
+			    	if(file.isDirectory())
+			    	{
+			    		File[] files = file.listFiles();
+			    		if(files.length>0)
+			    		{
+			    			files[0].delete();
+			    		}
+			    		file.delete();
+			    	}	
+			    	
+				   
 				   return;
 			   }
 			   
@@ -2190,10 +2235,10 @@ public class MmMapViewer extends JPanel implements Printable {
 			
 			for(int i=0;i<events.getStations().size();i++)
 		    {
-				 for(int j=0;j<events.getStations().get(i).getItemsCollected().size();j++)
+				 //for(int j=0;j<events.getStations().get(i).getItemsCollected().size();j++)
 				 {
-					  //JOptionPane.showMessageDialog(null, "objects collect "+events.getStations().get(i).getItemsCollected().get(j).intValue()+"   "+count);
-					  count+=events.getStations().get(i).getItemsCollected().get(j).intValue();
+					  //JOptionPane.showMessageDialog(null, "objects collect "+events.getStations().get(i).getItemsCollected().size()+"   "+count);
+					  count+=events.getStations().get(i).getItemsCollected().size();
 				 }
 		    }
 			
@@ -2201,11 +2246,15 @@ public class MmMapViewer extends JPanel implements Printable {
 			{
 				count+=markerEvents.get(i).getCollectedItems();
 			}
+			
+		    
 			   
-			if(count>1)   
+			/*if(count>1)   
 			   numCollectItemsAttributes.put("numCollectItems", count-1);
 			else
-				numCollectItemsAttributes.put("numCollectItems", count);	
+				numCollectItemsAttributes.put("numCollectItems", count);*/
+			
+			numCollectItemsAttributes.put("numCollectItems", count);
 			   
 			numCollectItems.put("attributes",numCollectItemsAttributes);
 			
@@ -2254,7 +2303,19 @@ public class MmMapViewer extends JPanel implements Printable {
 	       }
 	       catch(Exception e)
 	       {
-	    	   JOptionPane.showMessageDialog(null, "Kan inte spara json filen");
+	    	    JOptionPane.showMessageDialog(null, "Kan inte spara json filen");
+	    	   
+	    	    File file = new File(filePath);
+		    	if(file.isDirectory())
+		    	{
+		    		File[] files = file.listFiles();
+		    		if(files.length>0)
+		    		{
+		    			files[0].delete();
+		    		}
+		    		file.delete();
+		    	}	
+		    	return;
 	       }
 		
 	}
@@ -2267,7 +2328,7 @@ public class MmMapViewer extends JPanel implements Printable {
 		
 		JSONObject startName = new JSONObject();
 		JSONObject startAction = new JSONObject();
-		JSONArray startActionArray =  new JSONArray();
+		startActionArray =  new JSONArray();
 		
 		JSONObject numRegions = new JSONObject();
 		JSONObject numRegionsAttributes = new JSONObject();
@@ -2325,7 +2386,7 @@ public class MmMapViewer extends JPanel implements Printable {
 			   initializeEvents.put("name", "launch");
 			   initializeEvents.put("type", "generic");
 			   
-			   if(!this.getStations().isEmpty())
+			   if(!this.getStations().isEmpty() && (startEvents.getTexts().size()<=1))
 			       startActionArray.put("compassView");
 			   
 			   action.put("Start");
@@ -2352,7 +2413,7 @@ public class MmMapViewer extends JPanel implements Printable {
 		
 		if(!startEvents.getTexts().isEmpty())
 		{
-			for(int i=0;i<4;i++)
+			for(int i=1;i<2;i++)
 			{
 				if(!startEvents.eventNames[i].isEmpty())
 				{
@@ -2439,7 +2500,9 @@ public class MmMapViewer extends JPanel implements Printable {
 			{
 				//JOptionPane.showMessageDialog(null, globalMarkers.get(i).getMarkerName());
 				if(globalMarkers.get(i).getNumberOfEvents()!=0)
+				{
 					startActionArray.put(globalMarkers.get(i).getMarkerName());
+				}		
 			}
 		}	
 		
@@ -2632,7 +2695,7 @@ public class MmMapViewer extends JPanel implements Printable {
 			   {	   
 			       compassEvents.put("name", "enableStation"+Integer.toString(currentStation.getIndexStationNameIndex())+"Compass");
 			       if(currentStation.getCurrentStationIndex()==0)
-			    	   action.put("enableStation"+Integer.toString(currentStation.getIndexStationNameIndex())+"Compass");
+			    	   startActionArray.put("enableStation"+Integer.toString(currentStation.getIndexStationNameIndex())+"Compass");
 		   	       compassEvents.put("type", "compassmarker");
 			       compassAttributes.put("regionIdentifier", currentStation.getStationName());
 			       compassAttributes.put("showCompassMarker",true);
@@ -2667,7 +2730,7 @@ public class MmMapViewer extends JPanel implements Printable {
 		       {	   
 			       compassEvents1.put("name", "disableStation"+Integer.toString(currentStation.getIndexStationNameIndex())+"Compass");
 			       if(currentStation.getCurrentStationIndex()!=0)
-			    	   action.put("disableStation"+Integer.toString(currentStation.getIndexStationNameIndex())+"Compass");
+			    	   startActionArray.put("disableStation"+Integer.toString(currentStation.getIndexStationNameIndex())+"Compass");
 		   	       compassEvents1.put("type", "compassmarker");
 			       compassAttributes1.put("regionIdentifier", currentStation.getStationName());
 			       compassAttributes1.put("showCompassMarker",false);
@@ -2681,7 +2744,7 @@ public class MmMapViewer extends JPanel implements Printable {
 			       compassAttributes1.put("regionIdentifier", currentStation.getStationName());
 			       compassAttributes1.put("showCompassMarker",false);
 			       compassEvents1.put("attributes", compassAttributes1);
-			       action.put("disableStation"+Integer.toString(currentStation.getSwingPointStationIndex())+"SwingPoint"+Integer.toString(currentStation.getIndexStationNameIndex())+"Compass");
+			       startActionArray.put("disableStation"+Integer.toString(currentStation.getSwingPointStationIndex())+"SwingPoint"+Integer.toString(currentStation.getIndexStationNameIndex())+"Compass");
 			   }
 		    
 		} catch (JSONException e) {
@@ -2711,7 +2774,10 @@ public class MmMapViewer extends JPanel implements Printable {
 				object.put("attributes", attribute);
 				jsonObjects.put(object);
 				
-				action.put("mapMarker"+Integer.toString(i+1));
+				if(i==0)
+					action.put("mapMarker"+Integer.toString(i+1));
+				else
+					startActionArray.put("mapMarker"+Integer.toString(i+1));
 				
 				writeMarkerFiles("patt.map_marker"+Integer.toString(i+2),filePath);
 				
@@ -2814,10 +2880,10 @@ public class MmMapViewer extends JPanel implements Printable {
 		if(!writeFile("diamond.osg",filePath))
 			  return false;   
 		   
-		if(writeFile("base1.jpg",filePath))
+		if(!writeFile("base1.jpg",filePath))
 			return false;
 		
-		if(writeFile("diamond2.jpg",filePath))
+		if(!writeFile("diamond2.jpg",filePath))
 			return false;
 		
 		return true;
@@ -2926,17 +2992,17 @@ public class MmMapViewer extends JPanel implements Printable {
 	public void setSaved(boolean save)
 	{
 		isSaved = save;
-		events.setSaved(true);
-		mapMarkerWindow.setSaved(true);
+		events.setSaved(save);
+		mapMarkerWindow.setSaved(save);
 	}
 	
 	public boolean getSavedState()
 	{
 		
-		isSaved = events.getSavedState();
-		isSaved = mapMarkerWindow.getSavedState();
-		
-		return isSaved;
+		if(events.getSavedState() && mapMarkerWindow.getSavedState())
+			return true;
+			
+		return false;
 	}
 	
 	public void drawMapMarkers()
@@ -3989,6 +4055,10 @@ public class MmMapViewer extends JPanel implements Printable {
                 // convert from viewport to world bitmap
                 final Rectangle rect = mapKit.getMainMap().getViewportBounds();
                 
+                
+                
+                System.out.println("rect "+rect);
+                
                 //g.translate(-rect.x, -rect.y);
                 
                 if(!isPrintSelected)
@@ -4142,7 +4212,7 @@ public class MmMapViewer extends JPanel implements Printable {
                     
                     imageIndex = 2;
                     
-                    for (final GeoPosition gp : geoPos) {
+                    /*for (final GeoPosition gp : geoPos) {
                         // convert geo to world bitmap pixel
                     	
                     	                    	
@@ -4152,7 +4222,7 @@ public class MmMapViewer extends JPanel implements Printable {
                         g.drawString(Integer.toString(imageIndex-1)+":patt.map_marker"+Integer.toString(imageIndex), (int)(pt.getX()-rect.x-20),(int)(pt.getY()-rect.y+20));  
                         
                         imageIndex++;
-                    }
+                    }*/
                 }
                 
                             
@@ -4197,6 +4267,8 @@ public class MmMapViewer extends JPanel implements Printable {
     			pBook.append(this, pPageFormat);
     			printer.setPageable(pBook);
     			printer.print();
+    			
+    			
     		}
     		catch(Exception e)
     		{
@@ -4219,10 +4291,14 @@ public class MmMapViewer extends JPanel implements Printable {
        // by the X and Y values in the PageFormat
        // to avoid clipping.
        Graphics2D g2d = (Graphics2D)g;
+       g2d.setRenderingHint
+       (RenderingHints.KEY_ANTIALIASING, 
+         RenderingHints.VALUE_ANTIALIAS_ON);
        g2d.translate(print.getImageableX(), print.getImageableY());
+  
+      
 
-       mapKit.getMainMap().printAll(g);
-       
+       mapKit.getMainMap().printAll(g2d);
       
 
        // tell the caller that this page is part
